requirements.txt

Flask==0.10.1
gunicorn==19.4.5
numpy==1.16.3
pandas==0.24.2
scipy==1.2.1
python-Levenshtein-wheels
